Make the check for trailing characters smarter.
Although the `Time` class doesn't recognize timezones the `readFrom:` method needs to be able to accept a trailing 
timezone as `Time class >>> readFrom:` is used in `DateAndTime`, which does recognize timezones. We now peek
the stream to see if there are additional characters and if so check that they could be a timezone. If they could not be a
timezone, then an error is still thrown.