Improve the workaround to not show the  undeclared fix menu when in comment mode.
This PR is a workaround: as many classes can be used a requestor, adding a test method might lead to more problems, therefore the idea
is to detect that case when it is happening in comment mode.

Pharo10 will clean this up and not do the variable suggestion as part of the exception.

fixes #9560