It refactors the compiler by reifying OCSemanticScope's representing the meaning of the code: a method or a doIt expressions: - OCMethodSemanticScope - OCReceiverDoItSemanticScope - OCContextualDoItSemanticScope The configuration parameters of Compiler and CompilerContext #class, #context, #receiver, #noPattern are removed (but accessor methods are kept for compatibility). The code paths over them are replaced by delegation to a semantic scope.Thus the parsing and compilation methods are slightly simplified without ifNil/ifTrue checks spreaded in many places in the compiler. In future the existing users of configuration API (class: , receiver: ) needs to be cleaned to be based directly on semantic scopes instead of separate parameters. It will allow many interesting feature.
As a side effect of these changes (and it's driver actually) the OCContextualDoItSemanticScope brings a transparent access for temps in doIts expressions in the debugger.
Now the debugIt in the debugger shows real temp vars instead of reflective calls (ThisContext readVariableNamed:).
The reflective doIt with ThisContext is also kept with new name: #asReflectiveDoItForContext: . It is used by mustBeBoolean machinery. In future it can be simplified with special kind of DoItVariable.