Adding a new event that records changes in the display or resolutions. This is needed to correctly handle Remote Desktop connection and Disconnection.
Adding in the drawing of the renderer because we need to ensure that the error does not destroy everything.
The problem is if we have the exception in a mutex, and the mutex is not liberated.

Spliting the initialization of constants.
The Pharo 9 bootstrap is still building the image first in V3 bytecode and the number of literals is limited.
These methods can be inlined when V3 bytecode is left behind in the bootstrap.