Fix #10384
In Pharo 9, we have introduced a performance regression.
We have modified the implementation to use the implementation in Integer >> #/ when the primitive fails.
However, this implementation is intended to be used with a parameter of type LargeInteger.
The implementation in Integer uses a primitive implemented for LargeIntegers.
Passing a SmallInteger force the execution of the fallback code in the primitive.