The compiler has a lot of odd features as we wanted to be 100% backward compatible.
This PR simplifies the idea of #isInteractive:  if there is no requestor, we are not interactive.

This means we do not need the flag "interactive" on CompilationContext and can simplify.

For now we keep the odd feature that requestors can implmement #interative and return false, but don no thave to.
(I think we should simpliy that, too)