This adds an example and a test for Fluid Slot definitions: like with classes, the idea that we have to add all possible combinations of  setters leads to a combinatorial explosion. We need to be able to do things like
#slot => ExampleSlotWithFluidAPI  value1: #testvalue1;  value2: #testvalue2


if we ever want to be able to use this for anything non-trivial.

This PR just adds an example  on the level of Slots and a test that uses the class builder directly.

We can not define classes using this, but editing is broken due to the Fluid Class Parser enforcing "#name => ClassName oneSelector: #state", that is, the Fuid Class Parser enforces a non-fluid API for Slots, oddly. (see #6410)