- comment #allMethods to explain that it returns trait methods, too - add #methods that uses #localMethods
We do not use #methods anywhere, as #localMethods is from Traits and not avalaible in the bootstrap before Traits are installed.
This is no good, as just iterating over the methods that are local speeds clients that now use #allMethods.

issue #11928 stays open, but having #methods is important to be able to write scripts that analyse the system without having to explain traits implementation details.