This PR adds a setting #unrestrictedVariableDefinitions (default is off). Fixes #4629
The idea is that this allows easy experimentation, anything used in Pharo itself can not enable this option, but instead we would enhance the ClassParser
to parse the kind of defintions we want to support out of the box.

Here is the description text of the Setting:

EXPERIMENTAL: If true, the system will allow any expression as 
a Variable Definition (for Slots and Classvariables) that evaluates to a valid instance. 
		
WARNING1: you have to load all the Variable implementation classes *before* any users. 
WARNING2: the definition will be loaded by *unchecked* evaluation