Adds FinderPresenter class
Why: 
Making the first step to rewrite Finder in Spec2. 

How: 
Creates a new class named FinderPresenter, which is a subclass of SpPresenter. The class
provides a first layout for the new Finder. Internally it uses  the same search environment
in form of an RBBrowserEnvironment object as the old Finder.

Tags: Finder, Spec2