Adds FinderModel class
Why: 
The old model for the Finder tool is was not extensible and encapsulates multiple responsibilities in one class.
The new FinderModel class is much simpler and no longer implements the searches itself.

How: 
Creates a new class named FinderModel, which is a subclass of Model. The class
keeps track of the search environment and executes searches. 

Tags: Finder, Model