Add SearchTypesChanged Announcement
Why:
SearchTypesChanged is an announcement that will be made by the new Finder tool
to announce a change in the searches made available by it. The new Finder tool
allows to add and remove searches to it. This announcement signals a change to this
list.

How:
Creates a new class named SearchTypesChanged, which is a subclass of the ValueChanged
Announcement. It provides access to the new and old value.

Tags: Finder, Announcement