Connect FinderPresenter to FinderModel.
Why:
Uses FinderModel in FinderPresenter and connects the widgets to it. Now searches can be
performed via FinderPresenter and their results are displayed in a simple way.

How:
FinderPresenter is now a subclass of FinderPresenterWithModel.
This change implements updateResults and updateResultsWith: methods, which when called
update the resultTree with new data. updateSearchModes respectively updates the
searchModeDropList with the models availableSearches.
In subscribeToAnnouncements these update functions are used to called when the corresponding
announcements are made by the model.
Additionally the statusBar was disabled and left for future changes. connectPresenters and
searchInPackages got adjusted to call the right methods on the model.

Tags: Finder, Model, Presenter, Spec2