Add FinderResult class
Why:
A Dictionary was not particularly suited to be used as a result set as it did not provide a
uniform way to treat its contents and the SpTreePresenter is not made for displaying a
Dictionary. Thus FinderResult is implemented, which follows the composite pattern to
represent different results in a uniform way and nesting of FinderResult objects in a tree
like manner.

How:
FinderResult is the root composit class for results in the Finder tool. It implements a simple
tree like structure by having a parent and children. Additionally it implements the capabilities
for  actions and the actions for its content itsef. These actions correspond to the result
actions of the Finder tool.

Tags: Finder, Result