Add FinderSelectorResult class
Why:
FinderResult is the root composite object for results in the Finder tool. It provides
a uniform way to use results in the tool, but does not implement any specific actions
for a certian kind of result. FinderSelectorResult does this for selector results.

How:
FinderSelectorsResult subclasses FinderResult provides capabilities and action for
selectors.

Tags: Finder, Result