Dynamically enable resultButtons depending on selected result.
Why:
For a better user experience the result buttons now take the capabilities of the FinderResult's
into account and are enabled/disabled accordingly.

How:
Disable resultButtons by default. As soon as an item is selected in the resultTree widget, query
its capabilities and enable/disable the result buttons accordingly.

Tags: Finder, Presenter