Executable comments should manage string escaping
Let's imagine that we add an executable comments the should check that the result is `This is "exagerated" at least`. Since it is in a comment, we need to escape some caracters and it becomes `This is ""exagerated"" at least`. 
Currently the compared value did not remove the escaped double quotes.