Fix infinite loop in faulty literal array parsing
This is important for P11 because it can freeze the image on the parsing. 
The problem was that in the case of faulty literal array we never steped to look at the next character and we looped on the parsing of the faulty character of the faulty array.
I added a regression test also

Fixes #12886