Log the package been tested on CI.
The CI test runner is not always robust and it happens that we have crashes while running the tests. 
To make it a little easier to locate the problems, I propose here to write on wich package we are launching tests. Like this, if a crash happens, we can know which test package is responsible.