Unify package unregistration
We had basicUnregisterPackageNamed: and basicUnregisterPackage: to unregister packages but basicUnregisterPackageNamed: was missing two things:
- Logging multiple remove
- Unregistring MC working copy

I moved all the behavior in basicUnregisterPackageNamed: and removed the other method after fixing the two senders.
I'm not sure of the consequences but I'll let the CI show me if this is ok or not.

Fixes #13039