Simplify and deprecate ClassDescription>>allProtocolsUpTo:
allProtocolsUpTo: is not used in the image so I propose to deprecate it. 
In case an external tool is using it, I updated the method to be simpler so it can probably just be inlined. 

I added #nullProtocolName because Iâ¯renamed #nullCategory in another PR and I didn't want to do this PR in two steps.