Remove some unecessary not from the kernel
Sometimes we use not when a more readable way exist.

For example instead of having `(a > b) not` it is easier to read `a <= b` since it required only one step in out thoughts.
Or also, `self assert: a not` can just be `self deny: a`

This commit reduces the number of unecessary not in the kernel to help with readability.
Sorry if it reformat things but I did that with a script