Remove unused test ressources
I am not 100% sure that there are not needed in a test somewhere but let's see what the CI says.

I tried to browse the references, to look for the sources containing them and to iterate over all classes with shared pools and none of those methods returned any result.