Fix change of behavior
I don't like that we have magic in the compilation methods and opal instalation to not update the protocol in case it is unclassified. 
IMO we should not say "Compile into 'as yet unclassified'" but we should just give nils and deal with them at the end of the chain. Not carrying a useless protocol during all the chain