Make RPackageOrganizer polymorphic with SystemOrganizer
One goal of Pharo 12 is to make the package model part of the Pharo MM. Here is a step in that direction. 

If the package model is not really part of the MM yet is because the MM is maintaining a list of the packages tags and class names in SystemOrganizer and RPackage listen to the events of this class to update itself. 
Making the package model part of the MM means that we will remove the need of SystemOrganizer and make RPackage manipulate the classes directly. 

My first goal is to make Smalltalk hold the default RPackageOrganizer as an organization instead of the SystemOrganizer. Then when **only** RPackage will interact with it, we can inline the behavor and update it to manipulate real objects instead of names.