Remove references to SystemOrganization
SystemOrganization is a global variable holding the default SystemOrganizer. But this value can also be accessed via `Smalltalk organizer`. 
In order to avoid the multiple global variable, I am removing in the PR all references to SystemOrganization and in a future pull request I will kill this global variable and instead just save its instance un a normal instance variable of SystemDectionary, or better, SmalltalkImage if it is not to hard (because why the hell would it be the SystemDictionary that holds the PackageOrganizer?!)