Remove PBOldExtendedBootstrap
PBOldExtendedBootstrap does not seems to be referenced anywhere in the bootstrap process and the name indicated that it is something old. 
I propose to remove it to simplify the maintenance. 

When updating the packages and classes of Pharo core, we might have to update PBOldExtendedBootstrap>>#initializeClasses/initializeImage/initializeRPackages. But if this class is not used, then it simplify our tasks to just remove it and have a few less list of hardcoded classes/packages to maintain