Use date and time as Stdio windows file name distinguisher
Stdio on windows will sometimes create a new file and since we can have multiple ones, we were generating a random number in the file name.
Since the is the only dependency to Random-Core and that I would like to extract it from the KernelGroup of the bootstrap, I propose to use a timestamp instead.

I also added a test to be sure the dependency does not get added again.
One last change is the update of the manually resolved dependencies of Files package