Skip windows broken tests
The windows CI is ending up in timeout since the latest used VM because two tests and ending un OutOfMemory and the VM currently opens a native window blocking the execution of Pharo.

This skips those two tests until this issue is fixed (And it should get fixed before P11 release in my opinion because that means that first it is impossible to allocate 1Go on windows but also if that happens the execution get blocked. This is a show stopper for me

Fixes #13282