Remove the need to give a SystemOrganizer to SystemDictionary
SystemDictionary needs a SystemOrganizer to add or remove classes. Currently, people creating the SystemDictionary are creating this by hand.
This PR simplifies it by doing that during the initialization.

This also remove a bunch of other SystemOrganizer references a little everywhere in the system and the bootstrap which will help to replace the SystemOrganizer by a RPackageOrganizer.