Use Protocol>>#unclassified instead of #defaultName
We have two getter to have the unclassified protocol name. #unclassified and #defaultName.
This make the readability of the code harder and is useless since it is pretty obvious everywhere else in the code that by default we have an unclassified method.

I removed the need of #defaultName to use #unclassified everywhere.