Remove ClassOrganization>>#hasOrganizedClass
The announcement system of a class organization happens only if it has a class associated. But we should *always* have an associated class. 
Since the long terme goal is to clean ClassOraganization and move the remaining behavior to ClassDescription, I propose to remove all those checks directly because once it will be in ClassDescription it will be absolutly sure that we have the organized class.

Let's see if this will break something