Execute release tests last
We have a lot of tests in Pharo and it might happen that some of them do not respect the principle of letting the system clean with their tearDown.
To help to spot those impacts, I propose to run the releases tests last. Like this, if some bad generated code/or modifications of system entities happens, we might catch it with the verifications of release tests.

I expect this commit to break the tests in the CI but it'll help me find the culprits and fix them.