Rename Protocol class>>ambiguous
'ambiguous' is an ambiguous name in itself because it is hard to understand how it is useful with just the name.
I propose to rename it into Protocol class>>traitConflictName. 
Also, since it is an implementation detail of traits, I propose to move it to the TraitsV2 package.

I also did some renamings toward the goal of removing the "category" vocabulary.