Move class comment export to ClassDescription
A cleaning has started to move the comments managements from ClassOrganization to ClassDescription. 
This PR deals with the code export.

Remove ClassOrganization>>#putCommentOnFile:forClass: in favor of ClassDescription>>#putCommentOnFile:

I also added a print method to ClassOrganization to help with the refactors I'm doing and I merged two code import methods that I wanted to merge since a while.