Cut dependency from Collection-Streams to UIManager
Collection-Streams had 2 methods dealing with the UI which is a bad dependency IMO.
I could have added thos methods as an extention of the UI packages but they have no sender in the image. I propose to deprecate them. If a user wants to deal with a display bar he can do it himself.