Make protocol of super method a priority in MethodClassifier
MethodClassifier suggest to classify a method in the same protocol than its overriden method if one exist.  But this suggestion has the lowest priority of all.
Iâ¯suggest to give it the highest priority, especially since we have ReInconsistentMethodClassificationRule that yells at us if it is not the case.

I also adde a test to avoid an accidental reversal of the behavior.