Cut dependency from CommandLine and UIManager
My goal is to extract the UI manager from the KernelGroup build by espell in the bootstrap.
We have some magic in the command line to execute the command line with a high priority in interactive mode. Since there is no interactive mode before the UIManager is in the image, we can do this only after loading it.