Add logs to locate bug in protocols
Ideally, all methods of a class should always be in a protocol of the class, even if it is the unclassified protocol 
Sadly, some previous bootstrap crashes have shown that reality is harsh and this is not the case. This is why ClassOrganization>>#protocolNameOfElement:ifAbsent: comes with a hack that prevents the users to just use #protocolOfSelector: directly.

Let's add some logs to see if we can find the culprit and remove this ugly hack.