Fix random failure of Ombu in windows.
The problems comes from the fact that we use the current time to set the name of Ombu stores. 
But here we create 2 stores in less than 1 millisecond and DateAndTime on windows is sadly not able to give us a precision under 1 millisecond for now... See https://github.com/pharo-project/pharo/issues/13447.

So here is a quick hack to wait 1 millisecond before producing a new name on windows.