Introduce ClassDescription>>protocolOfSelector:
In the current system we have ClassOrganization>>protocolNameOfElement:ifAbsent:. that includes a hack when a method is present in the class but is not in its protocols (and some logs I added on the CI suggest that this is used a lot to get the protocol of a method in a superclass). In that case it returns Protocol unclassified.

I wan to remove totally the use of the method anyway to just manipulate protocols and not protocol names. If someone need the name they can just call #name of the protocol.

In order to do two birds one stone, I suggest the addtion of #protocolOfSelector on class description directly. This method would return the protocol of the method even if it is in a super class.
I might update this PR to start replacing the users of #protocolNameOfElement:ifAbsent: by this one