Use #protocolOfSelector: in CompiledMethod>>#category
I thought the method was storing its protocol but in fact it is asking the class for its protocol name using #protocolNameOfElement:

I want to remove this method because in case the method is not present in the class, it returns Protocol unclassified.
For me, if the method is not here, you return nil or an error. I'm going with the nil to simplify the transition currently. 

Since the change impacts a lot of places in the system, I will try to replace it little by little to see who was relying on the hack.
First place is CompiledMethod>>#category. Let's see if the CI agrees