Update RPackage to not silently rename protocols
When we rename a protocol we announce the change in the image so that some tools like Epicea can manage this change. 
Currently when renaming a RPackage, the extension renaming is not announced because it uses a silent rename.

I'm trying to avoid silent actions with protocol so let's see if anything breaks if we announce the change.