Fix bug in parenthesis highlight
In case we had parenthesis around a variable node, the style was wrong. 
This fixes the problem and adds a regression bug. 

I think we might have the same problem for other nodes too. 

I think we should add #visitValueNode: to the RB visitor so that we can manage it there directly for all nodes with parenthesis.