Make #renameProtocolNamed:toBe: manage nils
Currently there is a hack in the protocol management that returns Protocol unclassified as the protocol of some methods we cannot find the protocol of. 

I am hunting those, and in some cases the methods will say their protocol is nil if they are not in any protocol. To have a smooth transition, I'm making the APIâ¯deal with nils. This is not my favorite way, later I would like to have exceptions instead if possible. But I need to have a better understanding on how the system manipulates the protocols, so as a first step, I'm making the method silently cancel its action if one of the parameter is nil.