Fix bug in RPackage and SystemOrganizer synch
Currently we have two organization in the system managing the packages. 
- SystemOrganizer that is used by the system deeply
- RPackageOrganizer that is not really yet part of the Pharo MM

The goal is to make RPackageOrganizer part of the MM and to remove SystemOrganizer. But until we succeed (I hope during Pharo 12), we need to be sure that both are always in synch.
While improving the protocol management of Pharo I got a crash and while investigating on it, I found at least 3 synchronization problems. 

Here is a fix for the first one. If you send #removeFromSystem to a package of RPackageOragnizer, the package is unloaded but SystemOrganizer still has the categories corresponding in it.