Do not hardcode stylers
In the past Iâ¯had some personnal IconStyler and I noticed now that they stopped working. So reason is that in the past all stylers were applied. Now I found a hardcoded list of stylers in IconStyler class>>#withStaticStylers. 
This is really bad because it means that nobody can extend the styles. 

I now collect the stylers to apply in a way that allow to add your own.
I think we can do better but this already improves the situation