The management of protocol is currently hard to clean further because some hacks are present.
One of the hack is in #protocolNameOfElement:. This method check if an element is in the protocols of the class and if it is not, it will check if the method is included in the organized class. If it is, then it will return the unclassified protocol. 

One of the case where we rely on this hack is when we first compile a method. The method gets added to the method dictionary and then we classify it, meaning that it is not yet into any protocol.
Iâ¯would like to see what happens if we classify it before we add it to the method dictionary. Meaning that for the first compilation, the classification should have the behavior expected without the hack of #protocolNameOfElement: