Ensure methods are classified
All methods should be in a protocol when installed in a class. 
Currently, this is not the case. Some methods are added silently and never added to the class protocols. Then we have hacks that check if a method is classified or not in the protocol management.

Here I ensure that if a method is not classified while been added to a class, we put it in the unclassified protocol. 
If the method is installed by Opal or if the method was already classified in the past, then this will do nothing.