Simplify recompilation management
To install a method silently we have #addSelectorSilently:withMethod: and in case of recompilation we have #addSelector:withRecompiledMethod:
But if we take a look at #addSelector:withRecompiledMethod: then we can notice that it is useless.

It has 3 implementors:

1) In behavior

It just calls #addSelectorSilently:withMethod:

2) In TraitedClass

There it has a if. In the case the compile method is from a trait, it calls super (and thus #addSelectorSilently:withMethod:). 
In case the method does not come from a trait, it execute some code that is the same as the overriden version of #addSelectorSilently:withMethod:. So calling #addSelectorSilently:withMethod: would have the exact same effect. The only difference is the order of calls, but I'm not sure this has an impact.

In that case, we can probably just call #addSelectorSilently:withMethod:

3) In TraitedMetaclass

This is exactly the same as in TraitedClass.


I propose to remove all the recompilation methods to just use #addSelectorSilently:withMethod: instead. Let's see if the difference in the call orders has an impact.