Make shift tests more robust
Currently the shift test is removing from the system the generated classes. But a problem arise when the test fail during the class creation because the class was generated in the system but the variable was not set since an error happened before that. 
I propose to just generate the classes in a specific package that we remove in the tear down. Like this, even if the class creation fails, we can still remove the code and avoid getting the "Class XXX already exists" error