FIx case were method is in a class but not in a protocol
When a class has a slot, it generates an initialization method. But during the first creation of a class, this method got removed from its protocol when the class copied the organization of the old class.

This PR puts the organization copy before the initialization of slots to ensure the slot stay classified. It also adds a test.

This problem was detected by: #13533