First step at class name deprecation
Here is a first step at class name deprecation. 
We can now use the method ClassDescription>>deprecatedAliases: to set some deprecated aliases to a class. It will add a deprecated global variable to the SystemDictionary that can be referenced but that will be strikedthrought. 

Currently you can set this method in a class initialization.

I see a few next steps:
- Make Tonel export the information
- Add this to FluidClassBuilder so that we do not need to define it in the class initialization
- Warn users on accesses (This is for all deprecated global variables)
- Add a way to automatically rename the class references 

Fixes #13365