Do not reclassify already classified method
Currently the ClassOrganization has a guard to not reclassify a method if it is already in a protocol and the old and new protocol are the unclassified one. 
I don't know why we would do that only in the case a protocol is unclassified. IMO as soon as the old and new protocols are the same we should get in the guard clause.