Reduce cyclomatic complexity of #classify:under:
This is a new iteration on ClassOrganization>>#classify:under:. The goal is to manipulate a real protocol for the new protocol to be classified under to avoid an imbricated if.
There are still improvements to do to this method but since the changes to it seems to impact a lot the system I'm still doing small iterations to find what breaks and does not breaks the system and ultimatly find what hacks are currently been exploited by the system.