Cleanup protocol announcements in ClassOrganization
We had some methods to announce changes that did some guard to not announce things when there is no change. 
Now we should never get in those cases because we have guards to no go so far in the code in case the changes produces nothing. Thus we can remove the guards and inline the methods because they are really simple now. 
This reduces again the size of the ClassOrganization API to be able to inline it soon in ClassDescription