Test: Do not classify silently methods
We have two way of classifying methods in the system. One that announce the classification and one that does not. 
I found during the past few months places that were using the silent classification when they should have used the normal one, causing lost of informations in Epicea and other tools. 

We currently have only one last place with this silent classification: traits. 
I wonder if this is really needed. Here is a test to see if the image builds, if the tests passes and if the bootstrap timae changes with this change.