Fix: don't test unused transformations when they expect resumable `RBRefactoringError`
Transformations in general never raise `RBRefactoringWarning`.
This is a problem for parametrized tests because refactorings
do raise `RBRefactoringWarning` when a resumable error is thrown.

Since transformations only use `RBRefactoringError` some tests
became broken and fixing them means duplicating test code for
refactoring and changing expected error. Another thing is that both
of these transformations are unused in the image, I've removed
them from parametrized tests.

I don't think this is a problem, because we're starting to merging
refactorings and transformation to eliminate code and logic
duplication. This is a step in the right direction.