Use real protocol in recategorization mecanism
Now that the system is able to handle more and more Protocol instances instead of handling their names, we can use real protocol instance in the recategorization mecanisme avoiding some "protocol name" in multiple places.
This reduces the number of lookup to transform a name into a protocol while recategorizing.
In the code I updated we still use the name only in one place that is the MethodRecategorized announcement, but this is part of the next things to update. And this is a necessary step to simplify this future change since we now have our protocol instance in the method announcing the change instead of having a protocol name.