This is maybe a bit controversial:
PseudoVariables are just Variables, there is nothing that should, on the level of the class builder, raise a hard error if you want to shadow "self" with a slot named "self".

The tools will warn you that the variable is shadowing an existing variable of that name already, we could add a special rule (or improve the description) for this case in addition.